requirements.txt
fastapi
uvicorn
tensorflow==2.21.0
pillow
numpy
python-multipart